// src/auth/decorators/permissions.enum.ts
export enum Permission {
  // Role
  CreateRole = 'CreateRole',
  UpdateRole = 'UpdateRole',
  DeleteRole = 'DeleteRole',
  ViewRole = 'ViewRole',

  // Permission
  CreatePermission = 'CreatePermission',
  UpdatePermission = 'UpdatePermission',
  DeletePermission = 'DeletePermission',
  ViewPermission = 'ViewPermission',

  // User
  CreateUser = 'CreateUser',
  UpdateUser = 'UpdateUser',
  DeleteUser = 'DeleteUser',
  ViewUser = 'ViewUser',

  // Notification
  CreateNotification = 'CreateNotification',
  UpdateNotification = 'UpdateNotification',
  DeleteNotification = 'DeleteNotification',
  ViewNotification = 'ViewNotification',

  // Technicians
  CreateTechnicians = 'CreateTechnicians',
  UpdateTechnicians = 'UpdateTechnicians',
  DeleteTechnicians = 'DeleteTechnicians',
  ViewTechnicians = 'ViewTechnicians',

  // UserAddress
  CreateUserAddress = 'CreateUserAddress',
  UpdateUserAddress = 'UpdateUserAddress',
  DeleteUserAddress = 'DeleteUserAddress',
  ViewUserAddress = 'ViewUserAddress',

  // ServiceOrder
  CreateServiceOrder = 'CreateServiceOrder',
  UpdateServiceOrder = 'UpdateServiceOrder',
  DeleteServiceOrder = 'DeleteServiceOrder',
  ViewServiceOrder = 'ViewServiceOrder',

  // ServiceOrderDetail
  CreateServiceOrderDetail = 'CreateServiceOrderDetail',
  UpdateServiceOrderDetail = 'UpdateServiceOrderDetail',
  DeleteServiceOrderDetail = 'DeleteServiceOrderDetail',
  ViewServiceOrderDetail = 'ViewServiceOrderDetail',

  // Appointment
  CreateAppointment = 'CreateAppointment',
  UpdateAppointment = 'UpdateAppointment',
  DeleteAppointment = 'DeleteAppointment',
  ViewAppointment = 'ViewAppointment',

  // RepairImage
  CreateRepairImage = 'CreateRepairImage',
  UpdateRepairImage = 'UpdateRepairImage',
  DeleteRepairImage = 'DeleteRepairImage',
  ViewRepairImage = 'ViewRepairImage',

  // SupportRequests
  CreateSupportRequests = 'CreateSupportRequests',
  UpdateSupportRequests = 'UpdateSupportRequests',
  DeleteSupportRequests = 'DeleteSupportRequests',
  ViewSupportRequests = 'ViewSupportRequests',

  // Branch
  CreateBranch = 'CreateBranch',
  UpdateBranch = 'UpdateBranch',
  DeleteBranch = 'DeleteBranch',
  ViewBranch = 'ViewBranch',

  // ServiceCategory
  CreateServiceCategory = 'CreateServiceCategory',
  UpdateServiceCategory = 'UpdateServiceCategory',
  DeleteServiceCategory = 'DeleteServiceCategory',
  ViewServiceCategory = 'ViewServiceCategory',

  // Services
  CreateServices = 'CreateServices',
  UpdateServices = 'UpdateServices',
  DeleteServices = 'DeleteServices',
  ViewServices = 'ViewServices',

  // SparePartCategory
  CreateSparePartCategory = 'CreateSparePartCategory',
  UpdateSparePartCategory = 'UpdateSparePartCategory',
  DeleteSparePartCategory = 'DeleteSparePartCategory',
  ViewSparePartCategory = 'ViewSparePartCategory',

  // SparePart
  CreateSparePart = 'CreateSparePart',
  UpdateSparePart = 'UpdateSparePart',
  DeleteSparePart = 'DeleteSparePart',
  ViewSparePart = 'ViewSparePart',

  // Task
  CreateTask = 'CreateTask',
  UpdateTask = 'UpdateTask',
  DeleteTask = 'DeleteTask',
  ViewTask = 'ViewTask',

  // Review
  CreateReview = 'CreateReview',
  UpdateReview = 'UpdateReview',
  DeleteReview = 'DeleteReview',
  ViewReview = 'ViewReview',

  // Devices
  CreateDevices = 'CreateDevices',
  UpdateDevices = 'UpdateDevices',
  DeleteDevices = 'DeleteDevices',
  ViewDevices = 'ViewDevices',

  // UserNotification
  CreateUserNotification = 'CreateUserNotification',
  UpdateUserNotification = 'UpdateUserNotification',
  DeleteUserNotification = 'DeleteUserNotification',
  ViewUserNotification = 'ViewUserNotification',

  // ServiceWarranty
  CreateServiceWarranty = 'CreateServiceWarranty',
  UpdateServiceWarranty = 'UpdateServiceWarranty',
  DeleteServiceWarranty = 'DeleteServiceWarranty',
  ViewServiceWarranty = 'ViewServiceWarranty',
}